refactor: break down TransactionFlowDiagram into modular components
- Create TransactionFlowDiagram directory structure with separate files:
  - index.tsx: Main component file (cleaner, more focused)
  - hooks.ts: Custom useFlowPaths hook for SVG path calculations
  - utils.ts: Utility functions (buildCurvePath, formatAddress, formatScriptType, getScriptTypeColor, getStatusDisplay)
  - FlowDrawers.tsx: Input and output drawer components
  - FlowSummary.tsx: Summary section with legend and transaction totals
- Move formatSats and formatAddress to transactionFlowUtils.ts for reuse
- Remove old 1704-line monolithic TransactionFlowDiagram.tsx file

This addresses code review feedback to improve maintainability and readability.
